server/app/requirements.txt
faiss-cpu==1.11.0
numpy<2
fastapi[standard]
huggingface_hub
httpx
